# jk/speed up optimize ticks