- make S_view type stable
- fix boxing type-instability with `let`
- one more type instability

I timed with this:

```julia
@btime PlotUtils.optimize_ticks_typed(
    0.1123, 100.132, false,
    $([(1.0, 1.0), (5.0, 0.9), (2.0, 0.7), (2.5, 0.5), (3.0, 0.2)]),
    2,
    10,
    5,
    1 / 4,
    1 / 6,
    1 / 3,
    1 / 4,
    true,
    nothing,
    nothing,
)
```

And got the following timings:

| before | after |
| -- | -- |
| 5.406 ms (95897 allocations: 1.76 MiB) | 360.978 μs (23 allocations: 3.17 KiB) |

That's 15 times faster.

Should help quite a bit with the slowness that motivated this PR https://github.com/JuliaPlots/Makie.jl/pull/1612

Makes zooming in an Axis go from very choppy to smooth on my macbook.