@t-bltg 2022-03-14
Superseded by https://github.com/JuliaPlots/PlotUtils.jl/pull/136.